YDA-4407: fix intake movetovault locks
This re-enables locking on intake module movetovault jobs, so that
at most one movetovault job runs at any given time for a particular intake
group and the jobs can't interfere. An additional mechanism for
clearing intake group folder locks in case a movetovault job is
interrupted due to a transient failure has also been added.

An increase in the lock timeouts in the main ruleset is also part
of the fix for the issue.