pgbouncer: switch to direct MD5 authentication
The previous auth_user config required to set a password on the
Postgres account for authentication. Since iRODS uses only a single
database account, we can simplify this and just add manual
authentication for the iCAT database account instead. This way,
enabling or disabling PGBouncer does not affect the PostgreSQL
configuration.