YDA-5827: ensure securecopy present on consumer
In environments with a separate consumer server, we need to ensure
that the securecopy.sh script is present on the consumer as well,
since the publication code ultimately uses msiExecCmd to resolve
landing page and combi JSON data objects to a local file. If it resolves
one of these data objects to a replica on the consumer, it will try to run
securecopy from there.

In order to be able to run the securecopy script, we also need to install the
SSH key for copying files to the public server.