YDA-5882 fix ordering Yoda DB indexes
Deploying a new environment with Yoda-specific database indexes enabled
failed, because the playbook tried to create the indexes before iRODS
created the tables that were to be indexed.

This fix moves the code for enabling or disabling the Yoda-specific
indexes to a separate role that is always run after both the database
and the provider have been deployed, so that this doesn't happen anymore.