YDA-6045: fix Apache global TLS configuration
The Apache global TLS configuration was not applied on Ubuntu 20.04 LTS,
which resulted in the session cache not being created.

This change activates the global TLS configuration, including the
configuration of the session cache, so that we have session resumption.
We also needed to re-arrange Listen statements for TCP ports to ensure
each port has at most one Listen statement