YDA-6661: Move Flask session data to dedicated dir
Move EUS and portal Flask session data to dedicated
directory outside the /tmp filesystem.

The Apache systemd unit files on Ubuntu and EL have PrivateTmp
enabled, so if we store session data in /tmp, the web server
loses its session data every time the process is restarted or
(on EL) reloaded. This can cause user sessions to be terminated
unintentionally, e.g. when logs are rotated.